requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# mysql-crash/crash_test.py: 3,4
mysql_connector_python == 8.0.25

# mysql-crash/crash_test.py: 8
python_dotenv == 0.17.1
